Move Queue to App\Slave namespace
